requirements.txt
flask
azure-functions 